requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
appdirs==1.4.4
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
black==21.12b0
bleach==1.5.0
blessed==1.19.0
boto3==1.18.47
botocore==1.21.47
bravado==11.0.3
bravado-core==5.17.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
constantly==15.1.0
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.22
# dataclasses==0.8
decorator==4.4.2
Deprecated==1.2.13
flyingcircus==0.1.4.0
flyingcircus-numeric==0.1.1.2
freetype-py==2.2.0
future==0.18.2
gast==0.5.3
gitdb==4.0.7
GitPython==3.1.18
glcontext==2.3.4
grpcio==1.42.0
h5py==3.1.0
html5lib==0.9999999
hyperlink==21.0.0
idna==3.2
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.2
incremental==21.3.0
jmespath==0.10.0
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
mesh-to-sdf==0.0.14
meshio==4.4.6
moderngl==5.6.4
monotonic==1.6
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.2
multidict==5.2.0
mypy-extensions==0.4.3
neptune-client==0.11.0
networkx==2.5.1
nibabel==3.2.1
numeral==0.1.0.17
numex==0.0.0.7
numpy==1.23.0
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
plyfile==0.7.2
proglog==0.1.9
protobuf==3.19.1
psutil==5.8.0
pycparser==2.20
pyglet==1.5.15
PyGLM==2.5.1
PyJWT==2.1.0
pykdtree==1.3.4
pymadcad==0.8.4
PyMCubes==0.1.2
# pymesh2==0.3
pymeshlab==2021.10
pymrt==0.0.3.4
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyransac3d==0.6.0
pyrender==0.1.45
pyrsistent==0.18.0
python-dateutil==2.8.2
pytk==0.0.2.1
pytz==2021.1
pyvista==0.32.1
PyWavelets==1.1.1
PyYAML==5.4.1
raster-geometry==0.1.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
Rtree==0.9.7
s3transfer==0.5.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scooby==0.5.7
seaborn==0.11.2
setuptools-scm==6.3.2
simplejson==3.17.5
six==1.16.0
sklearn==0.0
smmap==4.0.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
sympy==1.9
# tensorboard==1.8.0
# tensorflow-gpu==1.8.0
# tensorflow-tensorboard==1.5.1
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2020.9.3
tomli==1.2.2
torch==1.12.0
tqdm==4.61.2
trimesh==3.9.21
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.7
vedo==2021.0.6
vtk==9.0.3
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.1
wslink==1.0.7
yarl==1.7.0
zipp==3.6.0
zope.interface==5.4.0
